setup integration tests for continous integration
