+++ tests/Titanium.Web.Proxy.IntegrationTests/InterceptionTests.cs
@@ -1,9 +1,10 @@
 ﻿using System;
 using System.Net;
 using System.Net.Http;
-using System.Net.Mime;
-using System.Text;
+using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Titanium.Web.Proxy.EventArguments;
+using Titanium.Web.Proxy.Models;
 
 namespace Titanium.Web.Proxy.IntegrationTests
 {
@@ -11,36 +12,64 @@ namespace Titanium.Web.Proxy.IntegrationTests
     public class InterceptionTests
     {
         [TestMethod]
-        public void CanInterceptPostRequests()
+        public async Task Can_Intercept_Post_Requests()
         {
             string testUrl = "http://interceptthis.com";
             int proxyPort = 8086;
             var proxy = new ProxyTestController();
             proxy.StartProxy(proxyPort);
 
-            using (var client = CreateHttpClient(testUrl, proxyPort))
+            using (var client = TestHelper.CreateHttpClient(testUrl, proxyPort))
             {
-                var response = client.PostAsync(new Uri(testUrl), new StringContent("hello!")).Result;
+                var response = await client.PostAsync(new Uri(testUrl), new StringContent("hello!"));
 
                 Assert.AreEqual(HttpStatusCode.OK, response.StatusCode);
 
-                var body = response.Content.ReadAsStringAsync().Result;
+                var body = await response.Content.ReadAsStringAsync();
                 Assert.IsTrue(body.Contains("TitaniumWebProxy-Stopped!!"));
             }
         }
 
-        private HttpClient CreateHttpClient(string url, int localProxyPort)
+        private class ProxyTestController
         {
-            var handler = new HttpClientHandler
+            private readonly ProxyServer proxyServer;
+
+            public ProxyTestController()
             {
-                Proxy = new WebProxy($"http://localhost:{localProxyPort}", false),
-                UseProxy = true
-            };
+                proxyServer = new ProxyServer();
+                proxyServer.CertificateManager.RootCertificateName = "root-certificate-for-integration-test.pfx";
+            }
 
-            var client = new HttpClient(handler);
+            public void StartProxy(int proxyPort)
+            {
+                proxyServer.BeforeRequest += OnRequest;
+                var explicitEndPoint = new ExplicitProxyEndPoint(IPAddress.Any, proxyPort, true);
+                proxyServer.AddEndPoint(explicitEndPoint);
+                proxyServer.Start();
+            }
 
-            return client;
+            public void Stop()
+            {
+                proxyServer.BeforeRequest -= OnRequest;
+                proxyServer.Stop();
+                proxyServer.Dispose();
+            }
+
+            public async Task OnRequest(object sender, SessionEventArgs e)
+            {
+                if (e.WebSession.Request.Url.Contains("interceptthis.com"))
+                {
+                    if (e.WebSession.Request.HasBody)
+                    {
+                        var body = await e.GetRequestBodyAsString();
+                    }
+
+                    e.Ok("<html><body>TitaniumWebProxy-Stopped!!</body></html>");
+                    return;
+                }
+
+                await Task.FromResult(0);
+            }
         }
-
     }
 }
